Crate arduino_uno

Board Support Crate for Arduino Uno (and compatible boards; see below).

This crate provides abstractions for interfacing with the hardware of Arduino Uno. It re-exports functionality from the underlying HAL in ways that make more sense for this particular board. For example, the pins are named by what is printed on the PCB instead of the MCU names.

Examples

A number of examples can be found in the examples/ subdirectory of this crate.

Getting Started

Please follow the guide from avr-hal's README for steps on how to set up a project with this board. A rough skeleton for an application looks like this:

#![no_std]
#![no_main]

// Pull in the panic handler from panic-halt
extern crate panic_halt;

// The prelude just exports all HAL traits anonymously which makes
// all trait methods available.  This is probably something that
// should always be added.
use arduino_uno::prelude::*;

// Define the entry-point for the application.  This can only be
// done once in the entire dependency tree.
#[arduino_uno::entry]
fn main() -> ! {
    // Get the peripheral singletons for interacting with them.
    let dp = arduino_uno::Peripherals::take().unwrap();

    unimplemented!()
}

Compatible Boards

This crate primarily targets Arduino Uno but is also compatible with a number of very similar boards. Where there are differences, those can be enabled with a crate feature. Here is an overview:

Feature	Board	Differences against Uno

arduino-nano	Arduino Nano	Additional ADC6 and ADC7 pins/channels (see examples/uno-adc.rs)

Re-exports

pub use atmega328p_hal as hal;

Modules

adc
Support for the Analog to Digital Converter

pac
ATmega328P

prelude

pwm
Support for PWM pins

spi
Support for the Serial Peripheral Interface

wdt
Support for the WatchDog Timer

Structs

DDR
Generic DDR that works for all ports

Peripherals
All the peripherals

Pins
Reexport of the Uno's pins, with the names they have on the board

Functions

delay_ms
Wait (busy spin) for ms milliseconds

delay_us
Wait (busy spin) for us microseconds

Type Definitions

Delay
Busy-Delay

I2cMaster
I2C Master on pins A4 (SDA) and A5 (SCL)

Serial
Serial (UART) interface on pins D0 (RX) and D1 (TX)

Attribute Macros

entry
See avr_device::entry.