Module arduino_uno::wdt

Support for the WatchDog Timer

Note

Changing the watchdog configuration requires two separate writes to WDTCSR where the second write must occur within 4 cycles of the first or the configuration will not change. You may need to adjust optimization settings to prevent other operations from being emitted between these two writes.

Example

let mut watchdog = arduino_uno::wdt::Wdt::new(&dp.CPU.mcusr, dp.WDT);
watchdog.start(arduino_uno::wdt::Timeout::Ms8000);

loop {
    watchdog.feed();
}

Structs

Wdt
Provides a system reset when a counter reaches a given time-out value.

Enums

Timeout
Approximate length of the time-out period before the watchdog provides a system reset.